backend/pyproject.toml
[tool.poetry]
name = "en-cavale"
version = "0.1.0"
description = ""
authors = ["Victor Nguyen <victong2@gmail.com>"]
readme = "README.md"
packages = [{include = "en_cavale"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg = "^3.2.1"
plotly = "^5.24.0"
pandas = "^2.2.2"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
psycopg2 = "^2.9.9"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
pandas-stubs = "^2.2.2.240909"
types-flask-migrate = "^4.0.0.20240311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
